Move Form types to Form container component
