+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Form/tests/Field.test.js
@@ -303,7 +303,7 @@ test('Merge with options from fieldRegistry before passing props to FieldType', 
         maxOccurs: 4,
         minOccurs: 2,
         options: {
-            anotherOption: 'anotherValue',
+            anotherOption: {value: 'anotherValue'},
         },
         type: 'text_line',
         types: {},
@@ -330,7 +330,7 @@ test('Merge with options from fieldRegistry before passing props to FieldType', 
         maxOccurs: 4,
         minOccurs: 2,
         schemaOptions: {
-            anotherOption: 'anotherValue',
+            anotherOption: {value: 'anotherValue'},
         },
         showAllErrors: true,
         types: {},
